delete old class PasswordChangeModal
