+++ packages/client/src/views/Settings/items/Password.tsx
@@ -24,7 +24,6 @@ import {
 } from '@client/i18n/messages'
 import { PasswordChangeModal } from '@client/views/Settings/PasswordChangeModal'
 import { useOnlineStatus } from '@client/utils'
-import { PasswordChangeModals } from '@client/views/Settings/PasswordChangeModals'
 
 export function Password() {
   const intl = useIntl()
@@ -65,12 +64,7 @@ export function Password() {
           </DynamicHeightLinkButton>
         }
       />
-      {/* <PasswordChangeModal
-        togglePasswordChangeModal={togglePasswordChangeModal}
-        showPasswordChange={showModal}
-        passwordChanged={changePassword}
-      /> */}
-      <PasswordChangeModals
+      <PasswordChangeModal
         togglePasswordChangeModal={togglePasswordChangeModal}
         showPasswordChange={showModal}
         passwordChanged={changePassword}
